+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     -   Moved `ButtonRoot` from `@lumx/react`
     -   Moved `Button` from `@lumx/react`
     -   Moved `ButtonGroup` from `@lumx/react`
+    -   Moved `IconButton` from `@lumx/react`
 
 ## [4.1.0][] - 2026-01-22